requirements.txt
sqlalchemy>=2.0
httpx>=0.25
python-dotenv>=1.0
streamlit>=1.28
pandas>=2.0
openpyxl>=3.1
plotly>=5.18
